% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCopepods.R
\name{getCopepods}
\alias{getCopepods}
\title{Find copepod occurrence data for given copepod species.
Given a copepod family, genus or species returns a list of copepod
occurrences.}
\usage{
getCopepods(species = NULL, genus = NULL, family = NULL, citation = TRUE)
}
\arguments{
\item{species}{Copepod species to filter}

\item{genus}{Copepod genus to filter}

\item{family}{Copepod family to filter}

\item{citation}{boolean. Should the output include the citation information?
default is FALSE}
}
\value{
A data.frame with copepod fish interactions
}
\description{
Find copepod occurrence data for given copepod species.
Given a copepod family, genus or species returns a list of copepod
occurrences.
}
\examples{
getCopepods(genus = "Caligus")
}
