% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFish.R
\name{getFish}
\alias{getFish}
\title{Find copepod occurrence data for given host fish.
Given a fish family, genus or species returns a list of copepod
occurrences on that host}
\usage{
getFish(species = NULL, genus = NULL, family = NULL, citation = TRUE)
}
\arguments{
\item{species}{Fish species to filter}

\item{genus}{Fish genus to filter}

\item{family}{Fish family to filter}

\item{citation}{boolean. Should the output include the citation information?
default is FALSE}
}
\value{
A data.frame with copepod fish interactions
}
\description{
Find copepod occurrence data for given host fish.
Given a fish family, genus or species returns a list of copepod
occurrences on that host
}
\examples{
getFish(genus = "Molva")
}
